Completed Main scene functionality
Created a base class "Shape", containing base variables and functions for all shapes. Attached to empty object in scene.

Created 3 child classes (Cube, Capsule, Cylinder) and attached to 3 3D objects in scene.

When clicked, each shape prints a message to the debug log and rotates a fixed amount. Most functionality is included in Shape class, though child classes can set variables declared in Shape, and there's an override method in Cylinder to display a custom message for that shape.

Added a get/set validator in Shape for ensuring the rotation amount variable (rotDegrees) is never set to a value less than 45.